projects/py-gateway/requirements.txt
pymysensors==0.23.0
